Disable encrypted private key support when in FIPS mode (#278)
